Move Neo4j vitest global setup to effect-vitest-testcontainers
